pyproject.toml
[tool.poetry]
name = "dcapy"
version = "0.1.1"
description = "Oil and Gas DCA Workflows"
authors = ["Santiago Cuervo"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.1"
numpy = "^1.20.1"
pandas = "^1.2.3"
matplotlib = "^3.3.4"
seaborn = "^0.11.1"
scipy = "^1.6.1"
pydantic = "^1.8.1"
numpy-financial = "^1.0.0"
PyYAML = "^5.4.1"
statsmodels = "^0.12.2"
openpyxl = "^3.0.7"
pyDOE2 = "^1.3.0"
rich = "^10.2.1"

[tool.poetry.dev-dependencies]
notebook = "^6.2.0"
ipykernel = "^5.5.0"
psycopg2-binary = "^2.8.6"
SQLAlchemy = "^1.3.23"
mkdocs-material = "^7.0.6"
mkdocstrings = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
